# openapi/FlorenceApi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "FlorenceApi",
    "description": "Self-hosted image recognition API powered by Microsoft Florence-2 (OpenVINO/Intel Arc). All requests are JSON with the image base64-encoded in the body. Authenticate by sending your key in the `X-API-Key` header.",
    "version": "v1"
  },
  "paths": {
    "/options": {
      "get": {
        "tags": [
          "Meta"
        ],
        "summary": "Runtime options reported by the worker (model, device, limits).",
        "description": "Returns the model id and Hugging Face revision currently loaded, the OpenVINO device\nin use (`GPU`, `GPU.0`, `CPU`, …), the max image upload size, and the accepted image\nformats. The list of recognition tasks lives in the OpenAPI spec at `/openapi/v1.json`\n— fetch that for client codegen.",
        "operationId": "GetOptions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionsResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ]
          }
        ]
      }
    },
...
    "/segmentations": {
      "post": {
        "tags": [
          "Segmentations"
        ],
        "summary": "Segment a region described by free text.",
        "description": "Given an image and a free-text expression (in `text`), returns one or more polygons\ncovering the matching region(s) plus labels.\n\n`polygons` is `polygons[i][j]` = polygon `j` of region `i`, each polygon a flat\n`[x1, y1, x2, y2, ...]` array of pixel coordinates. A region may consist of multiple\ndisjoint polygons (e.g. an object split by occlusion).\n\nMaps to Florence-2 task `<REFERRING_EXPRESSION_SEGMENTATION>`.",
        "operationId": "CreateSegmentation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SegmentationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentationResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "OptionsResponse": {
        "required": [
          "model",
          "revision",
          "device",
          "max_image_bytes",
          "supported_formats"
        ],
        "type": "object",
        "properties": {
          "model": {
            "type": "string"
          },
          "revision": {
            "type": "string"
          },
          "device": {
            "type": "string"
          },
          "max_image_bytes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "supported_formats": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "SegmentationRequest": {
        "required": [
          "text",
          "image"
        ],
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "media_type": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "SegmentationResult": {
        "required": [
          "polygons",
          "labels"
        ],
        "type": "object",
        "properties": {
          "polygons": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "array",
                "items": {
                  "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
                  "type": [
                    "number",
                    "string"
                  ],
                  "format": "double"
                }
              }
            }
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "API key. Send in the X-API-Key header.",
        "name": "X-API-Key",
        "in": "header"
      }
    }
  },
...
}